backend/requirements.txt
Flask
python-dotenv
langchain
langchain-community
langchain-google-genai
google-generativeai
chromadb

